Create explicit outputs in StackResources
When we are accessing attributes of resources in a nested stack (via
attributes of the parent resource), create output definitions in the nested
template to explicitly reference them.

In the long term, this should enable us to access the attribute values as
outputs directly using the APIs, rather than querying the resources
individually.

In the short term though, this means that access to these resource
attributes is no longer invisible to the child stack, so it will be able to
store the values of the attributes that are accessed (and ignore the ones
that are not).

Change-Id: I61d8e5deff45a648657ba685e033a00a492412f2
Partial-Bug: #1660831