Handle errors calculating dep_attrs for nested get_attr
Ensure that in convergence we can calculate the dep_attrs of a resource,
including the attributes referenced in outputs, even if:

- One of the resource's attributes is referenced in an ouput; and
- The attribute path itself also contains a get_attr function; and
- The attribute referenced by the inner get_attr function in the output is
  not referenced by any resources.

Since the attributes referenced only in outputs (and not in resources) are
not passed to dependent resources in the NodeData, the data required to
resolve the inner get_attr function is not available. Therefore getting the
dep_attrs for the relevant resource from the output definition would fail.
The next patch does so during the calculation of the NodeData, and any
resulting exception is not caught (bug 1703043), so the stack would end up
stuck IN_PROGRESS.

To avoid this, ignore any errors resolving attribute path components in the
course of calculating the dep_attrs of a given resource. This may mean that
an attribute value required for the outputs is not stored in the database,
but it can still be resolved live later.

Change-Id: I2ed1e3dacd371e36656559ba92dfada05df0ceba
Related-Bug: #1660831