Implement all_dep_attrs() more efficiently for get_attr
Override the default implementation of all_dep_attrs() (which calls
dep_attrs() in a loop). This makes all_dep_attrs() for GetAtt and its
arguments as efficient as dep_attrs(), even if we only need the data for a
single resource (as in convergence after we update a resource).

Change-Id: Ia6a4d132e655ef24fb158959d05967d7f1c7fef8
Related-Bug: #1684272