Cinder volume attachment attribute should not be cached
The Cinder volume 'attachments'attribute are
dynamically built lists and must be queried from the Cinder API on each
request. There are several factors that may change the contents of this
list that are unknown to Heat.  Caching should be prevented for this
attribute.

Closes-Bug: #1703605
Related-Bug: #1703387
Co-Authored-By: Paul Bourke <paul.bourke@oracle.com>
Change-Id: I1a646e567b2d88f2acf09b3077c205c5b061d4bc
(cherry picked from commit 9269787d39e052318ec59c58102cca128065b37a)