Handle PublicIp of None in AutoscalingGroup InstanceList
When the members of an InstanceGroup/AutoscalingGroup are nested stacks
(due to a mapping of AWS::EC2::Instance to a provider template in the
environment), getting the PublicIp attribute will return None instead of
the string '0.0.0.0' that AWS::EC2::Instance returns. Handle this case
without raising an exception.

The AWS AutoscalingGroup's version of the method is removed, since it is
identical to the parent class's.

Change-Id: I398fd985ca5df89d3d3eb145152132a2ec3055eb