Don't pass input_data to Resource.delete_convergence()
In the original prototype for convergence, we passed the input_data from
the SyncPoint to the resource when calling the equivalent of
convergence_delete(), so that we could clear and needed_by references that
no longer exist. This is pointless for a few reasons:

* It's implemented incorrectly - it *sets* the referenced resources into
  needed_by instead of clearing them from it.
* We don't actually pass any input data - in WorkerService.check_resource()
  it's always set to an empty dict for cleanup nodes, regardless of what
  came in on the wire.
* We don't store the result to the database unless we're deleting the
  resource anyway - in which case it doesn't matter.
* It turns out that even in the prototype, the whole needed_by mechanism
  isn't actually used for anything:
  https://github.com/zaneb/heat-convergence-prototype/commit/c74aac1f07e3fdf1fe382a7edce6c4828eda13e3

Rather than pretend that we're doing something useful with the input_data
here, just set the needed_by to an empty list, which is what was happening
anyway.

Change-Id: I73f6cf1646584dc4a83497f5a583c2c8973e8aba