Create replacement resources with correct requires
When we create a replacement resource, do so with the correct requires list
that it will ultimately have, instead of a copy of the old resource's
requires.

This will happen anyway when Resource.create_convergence() is actually
called (which is on the other end of an RPC message, so it may not actually
happen if another transition starts), but this makes it consistent from the
start.

Change-Id: Idf75a55be8d75e55c893ec1fb6ee3704f46bdc4f