Ignore exceptions getting FloatingIP dependencies
When calculating the dependencies for a FloatingIP resource, we make
REST calls to Neutron. If this call fails then we will stop looking for
implicit dependencies for that resource. (Prior to the fix for bug
1554625, it would have broken the stack altogether.)

This change ignores any errors that occur while making the client call,
so that any other implicit dependencies can potentially still be
discovered.

Change-Id: Ic9220041e7bf85b4388ad283082833cfbdb6c76a
Story: #1442121
Task: 16622