Extend addresses attribute of Zun container
The original 'addresses' field is a map from neutron network uuid
to an address list. For example:

  {'ed5c1b52-01ab-4a89-8b99-00705d9066c6': [..]}

This commit proposes to extend it to neutron network name. This
allows the address list to be hashed from uuid or name of a network.

  {'ed5c1b52-01ab-4a89-8b99-00705d9066c6': [...],
   'private': [...]}

This improvement will make the format of addresses consistent
between containers and VMs. In addition, it allows retrieving IP
addresses through either network name or network id. For example:

  {get_attr: [my-container, addresses, private, 0, addr]}

or

  {get_attr: [my-container, addresses, <uuid>, 0, addr]}

The first form is more user-friendly and the latter is able to
resolve ambiguity that two networks have the same name.

Closes-Bug: #1709191
Change-Id: Ib319e3683529fe71ee1c8dbe3087c9f30f1f1116