Switch integration jobs to use local mirrors
The integration jobs currently download images from a third party
mirror which can be slow at times or use up a lot of bandwidth.

This patch tries to automatically detect if nodepool mirrors are
active and uses them, falling back to the global round-robin mirror
of the Fedora project instead of a specific mirror.

Change-Id: Ie111c20b6dd0efe2e1cbfcac2a46b27ce6bc20c0