fix logger names
Most files are using getLogger(__name__) so that logging statements
are associated with the file they came from, but there are several
places where this practice has not been followed. This changes to
use __name__ consistently for most code. Per review comments,
commands and wsgi apps are instead updated to use the command/app
name.

Change-Id: Ic2594b88dd818df5486800e4759848d43edc24d1