Ignore errors in OS::stack_id output
If a provider stack contained an OS::stack_id output and there was an error
in the output, we would raise TemplateOutputError when trying to calculate
the reference ID of the facade resource. Since we do that in many API
calls, such an error could render the stack effectively unusable.

If we encounter such an error, log it and fall back to the default
reference ID.

Change-Id: I1bc921fe74c54eb0999541ef36afc42b9c19e9bc
Partial-Bug: #1712280
Related-Bug: #1719333