Avoid RPC call in TemplateResource.get_reference_id()
Most TemplateResources probably don't have an OS::stack_id output defined
in the template, so it's unfortunate that we have to make an RPC call to
check that every time we retrieve the resource's reference_id, which we
have to do for most stack operations.

To cut down on unnecessary RPC calls, check the template first locally to
see if there is an OS::stack_id output present.

Change-Id: Ia32ed6bca453b391371f544ad0a07d49dc0616e3
Related-Bug: #1719333