Fix get_watch_server_url for ipv6 addresses
https://review.openstack.org/#/c/508021 fixed it
to work in uwsgi mode but did broke it to work with
ipv6 as it contains ':' as part of the host.

This patch fixes how the URL is parsed in order to
work properly with ipv6.

Change-Id: I3965ee83e1f743e25cbdcbbec516be0ed6a30afc
Closes-Bug: #1721045
(cherry picked from commit 54ac107d7fee869c7cb8ed89f3e930f38ac48323)