Allow convergence-specific Tempest tests
Add a "convergence_engine_enabled" configuration option to the functional
tests' Tempest plugin. The option is enabled by default, but can be
disabled by setting the DISABLE_CONVERGENCE environment variable to "true"
when running prepare_test_env.sh, so that if convergence is disabled in
devstack it will also be disabled in Tempest.

This will allow us to write functional tests for convergence-specific (or
non-convergence-specific) behaviours.

Change-Id: If3a37de75467d50af10582215e16611e59a4ad06