Skip AodhAlarmTest
It seems ceilometer api has been removed[1].

As part od zuulv3 transition we are moving jobs to project repo[2]
and change to enable gnocchi devstack plugin at the gate has to wait
till the movement is complete.

We probably have no other option but to disable the test for the
time being.

[1] https://review.openstack.org/#/c/512286/
[2] https://review.openstack.org/#/c/509098/

Change-Id: I547b26ea78ad70742ae6fcf9a0eae4cd4af60100
Related-Bug: #1727637