Do deep redirects from old docs.openstack.org URLs
The docs.openstack.org site was rearranged so that the old docs at
/developer/heat/ are now at /heat/latest/. The base .htaccess file in
openstack-manuals redirects everything in the old hierarchy to the root
of the new one unless there is a .htaccess file present (see
I0af0e32d9c646d61f4c72599cdc4613299aa5ec9).

Add a .htaccess file so that old links redirect to the equivalent page
in the new location.

Change-Id: I74de09f4e25c2eda050053d3b8646586c9f24c04
Closes-Bug: #1722820