Document RBAC Policy allowed values
For entirely spurious reasons, RBAC Policies were validated using a custom
validate() method that is invisible to the user instead of using
constraints to validate values in a way consistent with all other resource
types, and that would appear in the API documentation. This rectifies that
error and improves the documentation and error reporting.

Change-Id: I0cde2fe5dd5c57c6f5a367d0405bcffad01678fa