Allow paths as lists in function.validate()
Previously when calling function.validate() we passed the path to the
function in the template (used for debugging purposes) as a string. This
change allows us to pass it as a list of path components, as we do
elsewhere, so that higher-level code that catches StackValidationFailed
can deal with its components separately.

Change-Id: I017aa6f7511b8478ef8273522ab8087684ae71c6