Change note for version/release from releasenotes
Release notes are version independent, so remove version/release values.
We've found that most projects now require the service package to be
installed in order to build release notes, and this is entirely due to
the current convention of pulling in the version information.

Release notes should not need installation in order to build, so this
unnecessary version setting needs to be removed.

This patch propose to change the note, so people won't try to use it.

Change-Id: I16110bb3951bd5bf9e644bbcd7b35b1f610ab909