Add more logging for software deployment updates
Include the server ID in existing logs, and add additional logs that will
let us know when updates to software deployments have happened and whether
they have succeeded or not. This should save considerable time
investigating in the future if there's ever any question of whether a
deployment has gone missing in Heat itself or somewhere on the target
server.

Change-Id: I33d7aee5974d03b7e41beab9fef4ad643cb0c469