Ignore resources with non-existent template
When listing resources with nested depth, we get all resources
for the root stack and cache them in the context. By the time
we iterate through the resources for a nested stack, if the
current_template_id of resources do not match the stack template,
we try to load the template. However, it's possible that the
template does not exist anymore. It would be good to ignore
those resources.

Change-Id: I838320539838ed74f490bca8601cde96eaf7c7ee
Closes-Bug: #1734815