Fix bypass list length check when all values are None
The patch https://review.openstack.org/#/c/527001/ bypassed the None
list but did not consider the case where all the for_each lists are
getting their values from attributes and therefore are None.

This patch set fix it by also skipping the check if all the values
are None -- as len(value_lens) would be 0 in that case.

Closes-bug: #1732934
Change-Id: Iba574bbd3877f9ca1ca755c384b3ddb47aec711c