Update openstacksdk construction
The queens release of python-openstacksdk breaks the sdk and senlin code
that's in heat right now. This shifts us to the new world order.

python-senlinclient is actually just a thin wrapper around
python-openstacksdk. Instead of reworking the constructor in the Senlin
plugin, just remove use of python-senlinclient and use the openstacksdk
client plugin instead.

Change-Id: Idf0acebf7b3774db26e335b3f3229227bfe68502