Fix attribute caching in Heat AutoScalingGroup
Attributes for an individual resource in an OS::Heat::AutoScalingGroup can
be retrieved using the attribute name 'resource.<index>'. Because we were
treating the index as a resource name when generating the template, we
would never generate an output in the nested template that referenced the
resource's attribute, so it wouldn't get cached. Since we can't know the
indices ahead of time, the only way to cache these is to do it for all
resources in the group - which is the same data cached by the equivalent
'outputs' attribute.

We also need to take into account both possible forms of syntax:

    [<asg_name>, resource.<index>.<attr_name>, ...]

or:

    [<asg_name>, resource.<index>, <attr_name>, ...]

Change-Id: I1d2898cdd4759b1bb9de1896a40056685e728f44
Closes-Bug: #1737047
Partial-Bug: #1731349
Related-Bug: #1660831