Get Instance/AutoScalingGroup attributes from nested stack outputs
Since Pike we've generated the templates for various nested stack resource
types with outputs that produce the attribute values that are referenced in
the parent stack, and a previous patch did the same for resource IDs. Use
these output values when available to calculate the attribute values. This
is efficient, because the all of the outputs are fetched together and
cached, and avoids using the grouputils functions that cause the nested
stack to get loaded into memory in the same process.

Fall back to the existing implementation (using grouputils) if the required
output isn't available (generally this would be due to the nested stack
being created on an earlier version of Heat that did not include the
outputs in the generated template, and not updated since).

The InstanceGroup and AutoScalingGroup implementations are more complex
than those of ResourceGroup and ResourceChain because they exclude failed
resources at runtime, and also must have a stable sort order for list-type
outputs that cannot be determined at template generation time. Therefore
the underlying output value is always a dict, and the attribute value is
calculated from it.

Change-Id: Ie4ad85068fdc16b8038a09df709ed7eb86434cfa
Partial-Bug: #1731349