Get ResourceGroup/Chain attributes from nested stack outputs
Since Pike we've generated the templates for various nested stack resource
types with outputs that produce the attribute values that are referenced in
the parent stack, and a previous patch did the same for resource IDs. Use
these output values when available to calculate the attribute values. This
is efficient, because the all of the outputs are fetched together and
cached, and avoids using the grouputils functions that cause the nested
stack to get loaded into memory in the same process.

Fall back to the existing implementation (using grouputils) if the required
output isn't available (generally this would be due to the nested stack
being created on an earlier version of Heat that did not include the
outputs in the generated template, and not updated since).

Change-Id: I9146e1d99f494213daef7d61ae75c92a4ef981a9
Partial-Bug: #1731349